[refactor] 은선님 pr 리뷰 반영
