+++ src/main/java/com/limito/payment/application/dto/resquest/CreatePaymentItemResquestV1.java
@@ -3,7 +3,6 @@ package com.limito.payment.application.dto.resquest;
 import java.util.List;
 import java.util.UUID;
 
-import com.limito.payment.infrastructure.jpa.entity.PaymentEntity;
 import com.limito.payment.domain.enums.PaymentStatusEnum;
 import com.limito.payment.presentation.dto.request.OrderItem;
 
@@ -25,5 +24,5 @@ public class CreatePaymentItemResquestV1 {
 	private int totalPrice;
 	@Size(min = 1, message = "결제 상품은 최소 1개 이상이어야 합니다.")
 	private List<OrderItem> items;
-	
+
 }
