--- src/main/java/com/limito/payment/domain/repository/PaymentItemRepository.java
@@ -1,4 +1,0 @@
-package com.limito.payment.domain.repository;
-
-public interface PaymentItemRepository {
-}
