+++ src/main/java/com/limito/payment/infrastructure/persistence/jpa/PaymentJpaRepository.java
@@ -1,0 +1,10 @@
+package com.limito.payment.infrastructure.persistence.jpa;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.limito.payment.infrastructure.persistence.entity.PaymentEntity;
+
+public interface PaymentJpaRepository extends JpaRepository<PaymentEntity, UUID> {
+}
