+++ src/main/java/com/limito/payment/presentation/PaymentControllerV1.java
@@ -9,9 +9,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import com.limito.payment.application.PaymentServiceV1;
 import com.limito.payment.presentation.dto.request.ConfirmPaymentRequestV1;
@@ -56,13 +54,8 @@ public class PaymentControllerV1 {
 	public ResponseEntity<ConfirmPaymentResponseV1> confirmPayment(
 		@PathVariable("paymentId") String paymentKey
 	) {
-		log.info("paymentKey={}", paymentKey);
-
 		ConfirmPaymentResponseV1 response = new ConfirmPaymentResponseV1();
-		response.setPaymentKey(paymentKey);
-
 		response = paymentService.confirmPayment(paymentKey, response);
 		return ResponseEntity.ok(response);
 	}
-
 }
